chore(weave): extract hardcoded_filters.py from calls_query_builder
Extract HardCodedFilter class and all process_*_filter_to_sql functions
into a dedicated hardcoded_filters.py module. These handle the direct
mapping from CallsFilter fields to optimized SQL WHERE conditions.

Kept process_children_of_eval_ids_to_sql in the query module since it
constructs CallsQuery instances (upward dependency).
